requirements.txt
fusepy>=3.0.0